requirements.txt
aws-cdk.core>=1.132.0
aws-cdk.assertions>=1.132.0
aws-cdk.aws_events>=1.132.0
aws-cdk.aws_events_targets>=1.132.0
aws_cdk.aws_logs>=1.132.0
boto3==1.16.7

